+++ shared/src/main/java/cz/vutbr/fit/group/verifit/oslc/shared/automationRequestExecution/RequestRunner.java
@@ -21,6 +21,7 @@ import java.nio.file.Path;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.Collection;
+import java.util.Map;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
@@ -115,10 +116,11 @@ public abstract class RequestRunner extends Thread {
 	 *                        means no time limit
 	 * @param id 			  Identifier that is appended to the stdout and stderr output file names (e.g. id="1" -> "stdout1", "stderr1")
 	 * @param filesToDeleteIfInterrupted	A collection of Files to be deleted in case the execution is interrupted. An output parameter.
+	 * @param envVariables	a map of name:value holding enviroment variables to be set when executing
 	 * @return a "ExecutionResult" object that holds the stdout, stderr, return code, timeout flag, etc (see the ExecutionResult class)
 	 * @throws InterruptedException 
 	 */
-	protected ExecutionResult executeString(Path folderPath, String stringToExecute, int timeout, String id, Collection<File> filesToDeleteIfInterrupted) throws InterruptedException
+	protected ExecutionResult executeString(Path folderPath, String stringToExecute, int timeout, String id, Collection<File> filesToDeleteIfInterrupted, Map<String,String> envVariables) throws InterruptedException
 	{
 		final String powershellExceptionExitCode = "1";
 		
@@ -170,6 +172,12 @@ public abstract class RequestRunner extends Thread {
 			File stderrFile = outputsDir.resolve("stderr" + id).toFile();
 			processBuilder.redirectOutput(stdoutFile);
 			processBuilder.redirectError(stderrFile);
+			
+			// set env variables
+			Map<String, String> processEnv = processBuilder.environment();
+			if (envVariables != null)
+				for (Map.Entry<String,String> envVar : envVariables.entrySet())
+					processEnv.put(envVar.getKey(), envVar.getValue());
 			
 		    // set files to be deleted in case of interrupt
 			filesToDeleteIfInterrupted.add(fileToExecute);